% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janno_basic.R
\name{as.janno}
\alias{as.janno}
\title{as.janno}
\usage{
as.janno(x, ...)
}
\arguments{
\item{x}{data.frame. Object to be converted to a janno object}

\item{...}{further arguments passed to or from other methods}
}
\description{
as.janno
}
