% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janno_calibrate.R
\name{calibrate}
\alias{calibrate}
\alias{calibrate.default}
\alias{calibrate.janno}
\title{Calibrate all valid dates in a \strong{janno} table}
\usage{
calibrate(x, ...)

\method{calibrate}{default}(x, ...)

\method{calibrate}{janno}(x, ...)
}
\arguments{
\item{x}{an object of class janno}

\item{...}{further arguments passed to or from other methods}
}
\description{
Calibrate all valid dates in a \strong{janno} table
}
