% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_module.R
\name{convert_module}
\alias{convert_module}
\title{Title}
\usage{
convert_module(output_format, input_package, output_directory, log_directory)
}
\arguments{
\item{output_format}{}

\item{input_package}{}

\item{output_directory}{}

\item{log_directory}{}
}
\description{
Title
}
