% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janno_enforce_types.R
\name{enforce_types}
\alias{enforce_types}
\alias{enforce_types.default}
\alias{enforce_types.janno}
\title{enforce_types}
\usage{
enforce_types(x, suppress_na_introduced_warnings = TRUE)

\method{enforce_types}{default}(x, suppress_na_introduced_warnings = TRUE)

\method{enforce_types}{janno}(x, suppress_na_introduced_warnings = TRUE)
}
\arguments{
\item{x}{object of class janno.}

\item{suppress_na_introduced_warnings}{suppress warnings caused by data removal in
type transformation due to wrong database entries (such as text in a number column)}
}
\description{
enforce_types
}
