% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/j2a.R
\name{j2a}
\alias{j2a}
\title{Title}
\usage{
j2a(
  in_janno_path,
  out_anno_path = file.path(dirname(in_janno_path), gsub(".janno", ".anno",
    basename(in_janno_path)))
)
}
\arguments{
\item{in_janno_path}{test}

\item{out_anno_path}{test}
}
\description{
Title
}
