% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janno.R
\docType{package}
\name{janno-package}
\alias{janno}
\alias{janno-package}
\title{janno: Poseidon .janno file handling in R}
\description{
Reading and manipulating Poseidon .janno files with a dedicated S3 data structure.
}
\author{
\strong{Maintainer}: Clemens Schmid \email{clemens@nevrome.de} (\href{https://orcid.org/0000-0003-3448-5715}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Ayshin Ghalichi [contributor]
}

}
\keyword{internal}
