% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janno_basic.R, R/janno_read.R,
%   R/janno_validate.R, R/janno_write.R
\name{janno}
\alias{janno}
\alias{as.janno}
\alias{format.janno}
\alias{print.janno}
\alias{read_janno}
\alias{validate_janno}
\alias{write_janno}
\alias{flatten_janno}
\title{\strong{janno}}
\usage{
as.janno(x, source_janno_file = NA_character_, ...)

\method{format}{janno}(x, ...)

\method{print}{janno}(x, only_header = FALSE, ...)

read_janno(path, to_janno = TRUE, validate = TRUE)

validate_janno(path)

write_janno(x, path, remove_source_file_column = TRUE)

flatten_janno(x, path, remove_source_file_column = TRUE)
}
\arguments{
\item{x}{an object}

\item{source_janno_file}{character. Path to the source .janno file.}

\item{...}{further arguments passed to or from other methods}

\item{only_header}{logical. Should only the header be printed.}

\item{path}{character vector. Paths to one or multiple .janno files or
directories that should be recursively searched for .janno files}

\item{to_janno}{logical. Should the read function transform the input
file to a janno object}

\item{validate}{logical. Run the file validation as part of the reading
process}

\item{remove_source_file_column}{Should the special column source_file be removed
from the janno object before writing?}
}
\description{
The \strong{janno} S3 class provides a data structure
derived from \link[tibble]{tibble} to represent the content of Poseidon
.janno files in R.
See the package README for more information.
}
\details{
\code{flatten_janno} transforms list columns to string lists (separated by ;)
}
