% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_module.R
\name{merge_module}
\alias{merge_module}
\title{Title}
\usage{
merge_module(input_file, output_directory, log_directory)
}
\arguments{
\item{input_file}{}

\item{output_directory}{}

\item{log_directory}{}
}
\description{
Title
}
