% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poseidonR.R
\docType{package}
\name{poseidonR-package}
\alias{poseidonR}
\alias{poseidonR-package}
\title{poseidonR: Poseidon data handling in R}
\description{
Interaction with Poseidon packages in R, mostly focused on .janno files.
}
\author{
\strong{Maintainer}: Clemens Schmid \email{clemens@nevrome.de} (\href{https://orcid.org/0000-0003-3448-5715}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Ayshin Ghalichi [contributor]
}

}
\keyword{internal}
