% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickcalibrate.R
\name{quickcalibrate}
\alias{quickcalibrate}
\title{quickcalibrate}
\usage{
quickcalibrate(uncalibrated_ages_BP, standard_deviations)
}
\arguments{
\item{uncalibrated_ages_BP}{character. One or multiple radiocarbon date ages BP
(e.g. "3000" or "3000,3100,3050" or c("2000", "2000,2300,2100"))}

\item{standard_deviations}{character. One or multiple standard deviations
(1 sigma ±) (e.g. "30" or "30,40,50" or c("20", "20,30,70"))}
}
\description{
Get values for the .janno file columns Date_BC_AD_Median,
Date_BC_AD_Start and Date_BC_AD_Stop from radiocarbon ages with mean
and sd.
}
