% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janno_upgrade.R
\name{upgrade_janno}
\alias{upgrade_janno}
\title{upgrade_janno}
\usage{
upgrade_janno(
  in_paths,
  out_paths,
  in_version = "2.4.0",
  out_version = "2.6.0",
  validate = TRUE
)
}
\arguments{
\item{in_paths}{character vector. Paths to one or multiple .janno files
or directories that should be recursively searched for .janno files}

\item{out_paths}{character vector. Must have an identical length to
\code{in_paths}, also considering the recursive search for .janno files
if directories are listed in \code{in_paths}. With
\code{poseidonR:::get_janno_file_paths(in_paths)} the output paths will match the
input paths, which will cause overwriting!}

\item{in_version}{character vector. Poseidon schema version number of the input files. Currently only supports "2.4.0"}

\item{out_version}{character vector. Poseidon schema version number of the output files. Currently only supports "2.5.0" and "2.6.0"}

\item{validate}{logical. Run the file validation after the transformation}
}
\value{
No output - writes files
}
\description{
A function to upgrade janno files from old versions of the Poseidon schema.
}
