% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_janno_file.R
\name{validate_janno_file}
\alias{validate_janno_file}
\title{validate_janno_file}
\usage{
validate_janno_file(path)
}
\arguments{
\item{path}{Character. Path to a .janno file}
}
\description{
Validate structural aspects of .janno files.
}
