% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_package_gitversion.R
\name{validate_packages_gitversion}
\alias{validate_packages_gitversion}
\title{validate_packages_gitversion}
\usage{
validate_packages_gitversion(package_search_path)
}
\arguments{
\item{input_packages}{Character vector. Paths to poseidon packages}
}
\description{
Special version of the package validation for the reduced representation
of the packages on github.
}
