% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a2j.R
\name{a2j}
\alias{a2j}
\title{Title}
\usage{
a2j(
  in_anno_path,
  out_janno_path = file.path(dirname(in_anno_path), gsub(".anno", ".janno",
    basename(in_anno_path)))
)
}
\arguments{
\item{in_anno_path}{test}

\item{out_janno_path}{test}
}
\description{
Title
}
