% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_module_convert.R,
%   R/cli_module_documentation.R, R/cli_module_extract.R, R/cli_module_merge.R,
%   R/cli_module_quickcalibrate.R, R/cli_module_validate.R
\name{convert_module}
\alias{convert_module}
\alias{cli_modules}
\alias{extract_module}
\alias{merge_module}
\alias{quickcalibrate_module}
\alias{validate_module}
\title{poseidon2 command line interface module functions}
\usage{
convert_module(
  output_format,
  input_package,
  output_directory,
  log_directory = tempdir()
)

extract_module(
  filter_file,
  input_package,
  output_directory,
  log_directory = tempdir()
)

merge_module(input_file, output_directory, log_directory = tempdir())

quickcalibrate_module(uncalibrated_ages_BP, standard_deviations)

validate_module(input_janno_file_or_packages)
}
\arguments{
\item{output_format}{character. Output format of conversion (e.g. eigenstrat)}

\item{input_package}{character. Path to input poseidon2 package}

\item{output_directory}{character. Path to output directory}

\item{log_directory}{character. Path to directory for log files and temporary data}

\item{filter_file}{character. Path to filter file}

\item{input_file}{character. Path to input file}

\item{uncalibrated_ages_BP}{character. One or multiple radiocarbon date ages BP (e.g. "3000" or "3000,3100,3050" or c("2000", "2000,2300,2100"))}

\item{standard_deviations}{character. One or multiple standard deviations (1 sigma ±) (e.g. "30" or "30,40,50" or c("20", "20,30,70"))}

\item{input_janno_file_or_packages}{character. Path to input janno file or poseidon2 package}
}
\description{
poseidon2 command line interface module functions
}
