% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_module.R, R/merge_module.R, R/module.R
\name{convert_module}
\alias{convert_module}
\alias{merge_module}
\alias{cli_modules}
\title{poseidon2 command line interface module functions}
\usage{
convert_module(output_format, input_package, output_directory, log_directory)

merge_module(input_file, output_directory, log_directory)
}
\arguments{
\item{output_format}{}

\item{input_package}{}

\item{output_directory}{}

\item{log_directory}{}

\item{input_file}{}
}
\description{
poseidon2 command line interface module functions
}
