% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R, R/extract.R, R/merge.R, R/module.R,
%   R/validate.R
\name{convert_module}
\alias{convert_module}
\alias{extract_module}
\alias{merge_module}
\alias{cli_modules}
\alias{validate_module}
\title{poseidon2 command line interface module functions}
\usage{
convert_module(
  output_format,
  input_package,
  output_directory,
  log_directory = tempdir()
)

extract_module(
  filter_file,
  input_package,
  output_directory,
  log_directory = tempdir()
)

merge_module(input_file, output_directory, log_directory = tempdir())

validate_module(input_janno_file_or_packages)
}
\arguments{
\item{output_format}{character. Output format of conversion (e.g. eigenstrat)}

\item{input_package}{character. Path to input poseidon2 package}

\item{output_directory}{character. Path to output directory}

\item{log_directory}{character. Path to directory for log files and temporary data}

\item{filter_file}{character. Path to filter file}

\item{input_file}{character. Path to input file}

\item{input_janno_file_or_packages}{character. Path to input janno file or poseidon2 package}
}
\description{
poseidon2 command line interface module functions
}
