% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janno_basic.R
\name{janno}
\alias{janno}
\alias{as.janno}
\alias{read_janno}
\alias{format.janno}
\alias{print.janno}
\title{\strong{janno}}
\usage{
as.janno(x, ...)

read_janno(
  file = "test_data/1_good_test_package/file1.janno",
  validate = TRUE,
  to_janno = TRUE
)

\method{format}{janno}(x, ...)

\method{print}{janno}(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{further arguments passed to or from other methods}

\item{file}{character. Path to a .janno file}

\item{validate}{logical. Should the janno file be validated upon reading}

\item{to_janno}{logical. Should the read function transform the input file to a janno object}
}
\description{
...
}
