% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janno_basic.R, R/validate_janno_file.R
\name{janno}
\alias{janno}
\alias{as.janno}
\alias{read_janno}
\alias{format.janno}
\alias{print.janno}
\alias{validate_janno}
\title{\strong{janno}}
\usage{
as.janno(x, ...)

read_janno(path, to_janno = TRUE, validate = TRUE)

\method{format}{janno}(x, ...)

\method{print}{janno}(x, only_header = FALSE, ...)

validate_janno(path)
}
\arguments{
\item{x}{an object}

\item{...}{further arguments passed to or from other methods}

\item{path}{character. Path to a .janno file or a directory that should
be recursively searched for .janno files}

\item{to_janno}{logical. Should the read function transform the input
file to a janno object}

\item{validate}{logical. Run the file validation as part of the reading
process}

\item{only_header}{logical. Should only the header be printed.}
}
\description{
The \strong{janno} S3 class provides a data structure
derived from \link[tibble]{tibble} to represent the content of poseidon
.janno files in R.
}
