% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janno_calibrate.R
\name{process_age}
\alias{process_age}
\alias{process_age.default}
\alias{process_age.janno}
\title{Process all valid dates in a \strong{janno} table}
\usage{
process_age(
  x,
  choices = c("Date_BC_AD_Prob", "Date_BC_AD_Median_Derived", "Date_BC_AD_Sample"),
  n = 100,
  ...
)

\method{process_age}{default}(
  x,
  choices = c("Date_BC_AD_Prob", "Date_BC_AD_Median_Derived", "Date_BC_AD_Sample"),
  n = 100,
  ...
)

\method{process_age}{janno}(
  x,
  choices = c("Date_BC_AD_Prob", "Date_BC_AD_Median_Derived", "Date_BC_AD_Sample"),
  n = 100,
  ...
)
}
\arguments{
\item{x}{an object of class janno}

\item{choices}{character vector. Which output columns should be added?}

\item{n}{integer. If "Date_BC_AD_Sample" in choices, then how many samples should be drawn? Default: 100}

\item{...}{further arguments passed to or from other methods}
}
\description{
Process all valid dates in a \strong{janno} table
}
