% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/janno_process_age.R
\name{process_age}
\alias{process_age}
\title{process_age}
\usage{
process_age(
  x,
  choices = c("Date_BC_AD_Prob", "Date_BC_AD_Start_Derived", "Date_BC_AD_Median_Derived",
    "Date_BC_AD_Stop_Derived", "Date_BC_AD_Sample"),
  n = 100,
  cal_curve = "intcal20",
  ...
)
}
\arguments{
\item{x}{an object of class \strong{janno}}

\item{choices}{character vector. Which output columns should be added?
By default all possible columns are added.}

\item{n}{integer. If "Date_BC_AD_Sample" is in \code{choices},
then how many samples should be drawn?}

\item{cal_curve}{String. Calibration curve to use for the (sum)calibration of
radiocarbon dates (see \link[Bchron]{BchronCalibrate})}

\item{...}{further arguments passed to or from other methods}
}
\description{
Process all valid dates in a \strong{janno} table.
}
