% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quickcalibrate.R
\name{quickcalibrate}
\alias{quickcalibrate}
\title{quickcalibrate}
\usage{
quickcalibrate(ages, sds)
}
\arguments{
\item{ages}{list. List of one or multiple radiocarbon date ages BP
(e.g. list(3000) or list(2000, c(2000, 2300, 2100)))}

\item{sds}{list. One or multiple standard deviations
(1 sigma ±) (e.g. list(30) or list(20, c(20, 30, 70)))}
}
\description{
Get values for the .janno file columns Date_BC_AD_Median,
Date_BC_AD_Start and Date_BC_AD_Stop from radiocarbon ages with mean
and sd.
See the package README for more information.
}
